Adopt for modern compiler

+++ tvision/classes/tdisplay.cc
@@ -381,10 +381,10 @@ Boolean TDisplay::searchClosestRes(TScreenResolution *res, unsigned x,
  if (firstXMatch!=-1)
    {// Return the closest y that match x
     i=indexMin=firstXMatch;
-    minDif=abs(res[i].y-y);
+    minDif=abs(static_cast<int>(res[i].y-y));
     while (++i<cant && res[i].x==x)
       {
-       dif=abs(res[i].y-y);
+       dif=abs(static_cast<int>(res[i].y-y));
        if (dif<minDif)
          {
           minDif=dif;
@@ -396,11 +396,11 @@ Boolean TDisplay::searchClosestRes(TScreenResolution *res, unsigned x,
    }
  // No x match, looks the one with minimum differences
  indexMin=0;
- minDif=abs(res[0].y-y)+abs(res[0].x-x);
+ minDif=abs(static_cast<int>(res[0].y-y))+abs(static_cast<int>(res[0].x-x));
  i=1;
  while (i<cant)
    {
-    dif=abs(res[i].y-y)+abs(res[i].x-x);
+    dif=abs(static_cast<int>(res[i].y-y))+abs(static_cast<int>(res[i].x-x));
     if (dif<minDif)
       {
        minDif=dif;
